tested Product class methods

+++ src/main/java/com/lab10/grupo5/CPS_20191_G5/Producto.java
@@ -2,8 +2,8 @@ package com.lab10.grupo5.CPS_20191_G5;
 
 public class Producto {
 
-	public double multlipicar(double a, double b) {
-		double res = a * b;
+	public int multlipicar(int a, int b) {
+		int res = a * b;
 		return res;
 	}
